requirements.txt
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==4.0.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.1
phonenumbers==8.13.13
Pillow==9.5.0
pyhtml2pdf==0.0.6
PyPDF2==3.0.1
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.2
urllib3==2.0.2
webdriver-manager==3.8.6
wsproto==1.2.0
